Move student app to the correct location
