+++ pom.xml
@@ -66,7 +66,7 @@
                         <id>default-cli</id>
                         <configuration>
                             <mainClass>
-                                org.roxburylatin.advcompsci.quizapp/org.roxburylatin.advcompsci.quizapp.application.HelloApplication
+                                org.roxburylatin.advcompsci.quizapp/org.roxburylatin.advcompsci.quizapp.application.student.StudentApplication
                             </mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
