+++ src/main/java/module-info.java
@@ -3,8 +3,8 @@ module org.roxburylatin.advcompsci.quizapp {
   requires javafx.fxml;
   requires org.jetbrains.annotations;
 
-  opens org.roxburylatin.advcompsci.quizapp.application to javafx.fxml;
-  exports org.roxburylatin.advcompsci.quizapp.application;
+  exports org.roxburylatin.advcompsci.quizapp.application.student;
+    opens org.roxburylatin.advcompsci.quizapp.application.student to javafx.fxml;
 
 //  exports org.roxburylatin.advcompsci.quizapp;
 }
