+++ src/main/java/org/roxburylatin/advcompsci/quizapp/application/student/StudentApplication.java
@@ -1,4 +1,4 @@
-package org.roxburylatin.advcompsci.quizapp.application;
+package org.roxburylatin.advcompsci.quizapp.application.student;
 
 import java.io.IOException;
 import javafx.application.Application;
@@ -6,14 +6,14 @@ import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.stage.Stage;
 
-public class HelloApplication extends Application {
+public class StudentApplication extends Application {
   public static void main(String[] args) {
     launch();
   }
 
   @Override
   public void start(Stage stage) throws IOException {
-    FXMLLoader fxmlLoader = new FXMLLoader(HelloApplication.class.getResource("hello-view.fxml"));
+    FXMLLoader fxmlLoader = new FXMLLoader(StudentApplication.class.getResource("student-question-view.fxml"));
     Scene scene = new Scene(fxmlLoader.load(), 320, 240);
     stage.setTitle("Hello!");
     stage.setScene(scene);
